backend/pyproject.toml
[tool.poetry]
name = "desafio-inoa"
version = "0.1.0"
description = ""
authors = ["Daniel Passy <danielpassy@hotmail.com>"]
readme = "README.md"
packages = [{include = "desafio_inoa"}]

[tool.poetry.dependencies]
python = "^3.10"
django-ninja = "^0.22.2"
ruff = "^0.0.279"
black = "^23.7.0"
pytest-django = "^4.5.2"
requests = "^2.31.0"
django-extensions = "^3.2.3"
ipython = "^8.14.0"
celery = "^5.3.1"
django-celery-results = "^2.5.1"
dj-database-url = "^2.0.0"
psycopg2 = "^2.9.6"
django-cors-headers = "^4.2.0"
sendgrid = "^6.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"
python_files = ["test_*.py", "*_tests.py", "tests.py"]
